# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IndiePact AI - Revenue Protection OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface RiskFix {
  rewrittenClause: string;
  direct: string;
  diplomatic: string;
  legal: string;
}

export type RiskSeverity = (typeof RiskSeverity)[keyof typeof RiskSeverity];

export const RiskSeverity = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface Risk {
  title: string;
  severity: RiskSeverity;
  explanation: string;
  whyThisHurtsYou: string;
  category: string;
  fixes: RiskFix;
}

export interface ScanResult {
  moneyImpactSummary: string;
  revenueAtRiskMin: number;
  revenueAtRiskMax: number;
  protectionScore: number;
  risks: Risk[];
  nextStep: string;
  rawExtractedClauses: string[];
}

export interface AnalyzeContractBody {
  /** @minLength 50 */
  contractText: string;
  userId: string;
  contractName?: string;
}

export interface SaveScanBody {
  userId: string;
  contractName: string;
  contractText: string;
  result: ScanResult;
}

export interface SavedScan {
  id: string;
  userId: string;
  contractName: string;
  contractText: string;
  result: ScanResult;
  createdAt: string;
  protectionScore: number;
  revenueAtRiskMin: number;
  revenueAtRiskMax: number;
  riskCount: number;
}

export interface ScansListResponse {
  scans: SavedScan[];
  total: number;
}

export interface DashboardSummary {
  totalScans: number;
  totalMoneyProtected: number;
  averageProtectionScore: number;
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  recentScans: SavedScan[];
}

export interface RiskTrendPoint {
  date: string;
  scopeCreep: number;
  paymentDelay: number;
  ipOwnership: number;
  liability: number;
  termination: number;
  revisionAbuse: number;
  vagueDeliverables: number;
}

export interface RiskTrendsResponse {
  trends: RiskTrendPoint[];
}

export interface ReportResponse {
  reportBase64: string;
  filename: string;
}

export type ListScansParams = {
  userId: string;
  limit?: number;
  offset?: number;
};

export type GetScanParams = {
  userId: string;
};

export type DeleteScanParams = {
  userId: string;
};

export type GetDashboardSummaryParams = {
  userId: string;
};

export type GetRiskTrendsParams = {
  userId: string;
  days?: number;
};

export type GenerateReportParams = {
  userId: string;
};